requirements.txt
Django==1.11.15
pytz==2018.7
xlrd==1.1.0
Pillow==5.3.0
django-summernote==0.8.11.1
pyopenssl
django-bootstrap4
docx-mailmerge

